// frontend/app/enums/colors.ts
import {Colors} from 'react-native/Libraries/NewAppScreen';

export const enum colors {
  primary = '#00EDAE',
  error = '#ff4d4d',
  warning = '#ffcc00',
  message = '#3399ff',
  lightContent = '#fff',
  darkContent = '#000',
  placeholderDark = '#5b5959',
  placeholderLight = '#a9a4a4',
  bgColorComponentsLight = '#dcdcdc',
  bgColorComponentsDark = '#343333',
}

export const darkModeBgColor = Colors.darker;
export const lightModeBgColor = Colors.lighter;